backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.1.0
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.6
redis==5.0.1
yfinance>=0.2.0
